remove the previous segmented control implementation
