+++ packages/lake/__stories__/SegmentedControl.stories.tsx
@@ -2,7 +2,6 @@ import { Meta } from "@storybook/react";
 import { useState } from "react";
 import { Icon } from "../src/components/Icon";
 import { SegmentedControl } from "../src/components/SegmentedControl";
-import { SegmentedControl2 } from "../src/components/SegmentedControl2";
 import { Path, Svg } from "../src/components/Svg";
 import { TabView } from "../src/components/TabView";
 import { StoryBlock, StoryPart } from "./_StoriesComponents";
@@ -45,7 +44,6 @@ export const Default = () => {
   const [selected, setSelected] = useState<ItemId>(items[0].id);
   const [noIconSelected, setNoIconSelected] = useState<ItemId>(items[0].id);
   const [fixedWidthSelected, setFixedWidthSelected] = useState<ItemId>(items[0].id);
-  const [flexWidthSelected, setFlexWidthSelected] = useState<ItemId>(items[0].id);
 
   return (
     <StoryBlock
@@ -64,16 +62,8 @@ export const Default = () => {
         />
       </StoryPart>
 
-      <StoryPart title="With default width">
-        <SegmentedControl2
-          selected={flexWidthSelected}
-          items={items}
-          onValueChange={setFlexWidthSelected}
-        />
-      </StoryPart>
-
-      <StoryPart title="With full width">
-        <SegmentedControl2
+      <StoryPart title="Full width">
+        <SegmentedControl
           selected={fixedWidthSelected}
           fullWidth={true}
           items={items}
